+++ frontend/__tests__/components/features/home/git-repo-dropdown.test.tsx
@@ -1,0 +1,15 @@
+import { test, expect } from "vitest";
+import { GitRepoDropdown } from "#/components/features/home/git-repo-dropdown";
+
+test("GitRepoDropdown should be importable", () => {
+  // Test that the component can be imported without errors
+  expect(GitRepoDropdown).toBeDefined();
+  expect(typeof GitRepoDropdown).toBe("function");
+});
+
+// TODO: Add comprehensive tests for GitRepoDropdown component
+// These tests would require proper mocking of:
+// - React Query hooks (useRepositories, useSearchRepositories)
+// - API endpoints for GitHub and GitLab
+// - Downshift behavior for dropdown interactions
+// - URL parsing and search functionality
