+++ frontend/src/components/features/home/git-repo-dropdown/clear-button.tsx
@@ -1,0 +1,40 @@
+import React from "react";
+import { cn } from "#/utils/utils";
+
+interface ClearButtonProps {
+  disabled: boolean;
+  onClear: () => void;
+}
+
+export function ClearButton({ disabled, onClear }: ClearButtonProps) {
+  return (
+    <button
+      onClick={(e) => {
+        e.stopPropagation();
+        onClear();
+      }}
+      disabled={disabled}
+      className={cn(
+        "p-1 text-gray-400 hover:text-gray-600",
+        "disabled:cursor-not-allowed"
+      )}
+      type="button"
+      aria-label="Clear selection"
+      data-testid="git-repo-dropdown-clear"
+    >
+      <svg
+        className="w-4 h-4"
+        fill="none"
+        stroke="currentColor"
+        viewBox="0 0 24 24"
+      >
+        <path
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          strokeWidth={2}
+          d="M6 18L18 6M6 6l12 12"
+        />
+      </svg>
+    </button>
+  );
+}
