+++ frontend/src/components/features/home/git-repo-dropdown/git-repo-dropdown.tsx
@@ -1,0 +1,217 @@
+import React, { useState, useMemo, useCallback, useRef, useEffect } from "react";
+import { useCombobox } from "downshift";
+import { Provider } from "#/types/settings";
+import { GitRepository } from "#/types/git";
+import { useDebounce } from "#/hooks/use-debounce";
+import { cn } from "#/utils/utils";
+import { LoadingSpinner } from "./loading-spinner";
+import { ClearButton } from "./clear-button";
+import { ToggleButton } from "./toggle-button";
+import { DropdownMenu } from "./dropdown-menu";
+import { ErrorMessage } from "./error-message";
+import { useUrlSearch } from "./use-url-search";
+import { useRepositoryData } from "./use-repository-data";
+
+export interface GitRepoDropdownProps {
+  provider: Provider;
+  value?: string | null;
+  placeholder?: string;
+  className?: string;
+  disabled?: boolean;
+  onChange?: (repository?: GitRepository) => void;
+}
+
+export function GitRepoDropdown({
+  provider,
+  value,
+  placeholder = "Search repositories...",
+  className,
+  disabled = false,
+  onChange,
+}: GitRepoDropdownProps) {
+  const [inputValue, setInputValue] = useState("");
+  const debouncedInputValue = useDebounce(inputValue, 300);
+  const menuRef = useRef<HTMLUListElement>(null);
+
+  // Process search input to handle URLs
+  const processedSearchInput = useMemo(() => {
+    if (debouncedInputValue.startsWith("https://")) {
+      const match = debouncedInputValue.match(
+        /https:\/\/[^/]+\/([^/]+\/[^/]+)/,
+      );
+      return match ? match[1] : debouncedInputValue;
+    }
+    return debouncedInputValue;
+  }, [debouncedInputValue]);
+
+  // URL search functionality
+  const { urlSearchResults, isUrlSearchLoading } = useUrlSearch(inputValue, provider);
+
+  // Repository data management
+  const {
+    repositories,
+    allRepositories,
+    selectedRepository,
+    fetchNextPage,
+    hasNextPage,
+    isLoading,
+    isFetchingNextPage,
+    isError,
+    isSearchLoading,
+  } = useRepositoryData(
+    provider,
+    disabled,
+    processedSearchInput,
+    urlSearchResults,
+    inputValue,
+    value
+  );
+
+  // Filter repositories based on input value
+  const filteredRepositories = useMemo(() => {
+    // If we have URL search results, show them directly (no filtering needed)
+    if (urlSearchResults.length > 0) {
+      return repositories;
+    }
+    
+    // If we have a selected repository and the input matches it exactly, show all repositories
+    if (selectedRepository && inputValue === selectedRepository.full_name) {
+      return repositories;
+    }
+    
+    // If no input value, show all repositories
+    if (!inputValue || !inputValue.trim()) {
+      return repositories;
+    }
+    
+    // For URL inputs, use the processed search input for filtering
+    const filterText = inputValue.startsWith("https://") ? processedSearchInput : inputValue;
+    
+    return repositories.filter((repo) =>
+      repo.full_name.toLowerCase().includes(filterText.toLowerCase())
+    );
+  }, [repositories, inputValue, selectedRepository, urlSearchResults, processedSearchInput]);
+
+  // Handle selection
+  const handleSelectionChange = useCallback(
+    (selectedItem: GitRepository | null) => {
+      onChange?.(selectedItem || undefined);
+      // Update input value to show selected item
+      if (selectedItem) {
+        setInputValue(selectedItem.full_name);
+      }
+    },
+    [onChange]
+  );
+
+  // Handle clear selection
+  const handleClear = useCallback(() => {
+    handleSelectionChange(null);
+    setInputValue("");
+  }, [handleSelectionChange]);
+
+  // Handle input value change
+  const handleInputValueChange = useCallback(
+    ({ inputValue: newInputValue }: { inputValue?: string }) => {
+      setInputValue(newInputValue || "");
+    },
+    []
+  );
+
+  // Handle scroll to bottom for pagination
+  const handleMenuScroll = useCallback(
+    (event: React.UIEvent<HTMLUListElement>) => {
+      const { scrollTop, scrollHeight, clientHeight } = event.currentTarget;
+      const isNearBottom = scrollTop + clientHeight >= scrollHeight - 10;
+      
+      if (isNearBottom && hasNextPage && !isFetchingNextPage && !isLoading) {
+        fetchNextPage();
+      }
+    },
+    [hasNextPage, isFetchingNextPage, isLoading, fetchNextPage]
+  );
+
+  const {
+    isOpen,
+    getToggleButtonProps,
+    getLabelProps,
+    getMenuProps,
+    getInputProps,
+    highlightedIndex,
+    getItemProps,
+    selectedItem,
+    reset,
+  } = useCombobox({
+    items: filteredRepositories,
+    itemToString: (item) => item?.full_name || "",
+    selectedItem: selectedRepository,
+    onSelectedItemChange: ({ selectedItem }) => {
+      handleSelectionChange(selectedItem);
+    },
+    onInputValueChange: handleInputValueChange,
+    inputValue,
+  });
+
+  // Initialize input value when selectedRepository changes (but not when user is typing)
+  useEffect(() => {
+    if (selectedRepository && !isOpen && inputValue !== selectedRepository.full_name) {
+      setInputValue(selectedRepository.full_name);
+    } else if (!selectedRepository && !isOpen && inputValue) {
+      setInputValue("");
+    }
+  }, [selectedRepository, isOpen]);
+
+  const isLoadingState = isLoading || isSearchLoading || isFetchingNextPage || isUrlSearchLoading;
+
+  return (
+    <div className={cn("relative", className)}>
+      <div className="relative">
+        <input
+          {...getInputProps({
+            disabled,
+            placeholder,
+            className: cn(
+              "w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm",
+              "focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500",
+              "disabled:bg-gray-100 disabled:cursor-not-allowed",
+              "pr-10" // Space for toggle button
+            ),
+          })}
+          data-testid="git-repo-dropdown"
+        />
+        
+        <div className="absolute right-2 top-1/2 transform -translate-y-1/2 flex items-center gap-1">
+          {selectedRepository && (
+            <ClearButton disabled={disabled} onClear={handleClear} />
+          )}
+          
+          <ToggleButton 
+            isOpen={isOpen} 
+            disabled={disabled} 
+            getToggleButtonProps={getToggleButtonProps} 
+          />
+        </div>
+
+        {isLoadingState && (
+          <LoadingSpinner selectedRepository={selectedRepository} />
+        )}
+      </div>
+
+      <DropdownMenu
+        isOpen={isOpen}
+        filteredRepositories={filteredRepositories}
+        isLoadingState={isLoadingState}
+        inputValue={inputValue}
+        highlightedIndex={highlightedIndex}
+        selectedItem={selectedItem}
+        isFetchingNextPage={isFetchingNextPage}
+        getMenuProps={getMenuProps}
+        getItemProps={getItemProps}
+        onScroll={handleMenuScroll}
+        menuRef={menuRef}
+      />
+
+      <ErrorMessage isError={isError} />
+    </div>
+  );
+}
