+++ frontend/src/components/features/home/git-repo-dropdown/index.tsx
@@ -1,0 +1,17 @@
+// Main component
+export { GitRepoDropdown } from "./git-repo-dropdown";
+export type { GitRepoDropdownProps } from "./git-repo-dropdown";
+
+// UI Components
+export { LoadingSpinner } from "./loading-spinner";
+export { ClearButton } from "./clear-button";
+export { ToggleButton } from "./toggle-button";
+export { RepositoryItem } from "./repository-item";
+export { EmptyState } from "./empty-state";
+export { LoadingMoreState } from "./loading-more-state";
+export { DropdownMenu } from "./dropdown-menu";
+export { ErrorMessage } from "./error-message";
+
+// Custom Hooks
+export { useUrlSearch } from "./use-url-search";
+export { useRepositoryData } from "./use-repository-data";
