+++ frontend/src/components/features/home/git-repo-dropdown/loading-more-state.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function LoadingMoreState() {
+  return (
+    <li className="px-3 py-2 text-center text-sm text-gray-500">
+      Loading more repositories...
+    </li>
+  );
+}
