+++ frontend/src/components/features/home/git-repo-dropdown/use-repository-data.tsx
@@ -1,0 +1,71 @@
+import { useMemo } from "react";
+import { Provider } from "#/types/settings";
+import { GitRepository } from "#/types/git";
+import { useGitRepositories } from "#/hooks/query/use-git-repositories";
+import { useSearchRepositories } from "#/hooks/query/use-search-repositories";
+
+export function useRepositoryData(
+  provider: Provider,
+  disabled: boolean,
+  processedSearchInput: string,
+  urlSearchResults: GitRepository[],
+  inputValue: string,
+  value?: string | null
+) {
+  // Fetch user repositories with pagination
+  const {
+    data: repoData,
+    fetchNextPage,
+    hasNextPage,
+    isLoading,
+    isFetchingNextPage,
+    isError,
+  } = useGitRepositories({
+    provider,
+    enabled: !disabled,
+  });
+
+  // Search repositories when user types
+  const { data: searchData, isLoading: isSearchLoading } =
+    useSearchRepositories(processedSearchInput, provider);
+
+  // Combine all repositories from paginated data
+  const allRepositories = useMemo(() => {
+    return repoData?.pages?.flatMap((page) => page.data) || [];
+  }, [repoData]);
+
+  // Find selected repository from all repositories
+  const selectedRepository = useMemo(() => {
+    return allRepositories.find((repo) => repo.id === value) || null;
+  }, [allRepositories, value]);
+
+  // Get repositories to display (URL search, regular search, or all repos)
+  const repositories = useMemo(() => {
+    // Prioritize URL search results when available
+    if (urlSearchResults.length > 0) {
+      return urlSearchResults;
+    }
+    
+    // Don't use search results if input exactly matches selected repository
+    const shouldUseSearch = processedSearchInput && 
+      searchData && 
+      !(selectedRepository && inputValue === selectedRepository.full_name);
+    
+    if (shouldUseSearch) {
+      return searchData;
+    }
+    return allRepositories;
+  }, [urlSearchResults, processedSearchInput, searchData, allRepositories, selectedRepository, inputValue]);
+
+  return {
+    repositories,
+    allRepositories,
+    selectedRepository,
+    fetchNextPage,
+    hasNextPage,
+    isLoading,
+    isFetchingNextPage,
+    isError,
+    isSearchLoading,
+  };
+}
