+++ frontend/src/components/features/home/git-repo-dropdown/use-url-search.tsx
@@ -1,0 +1,42 @@
+import { useState, useEffect } from "react";
+import { Provider } from "#/types/settings";
+import { GitRepository } from "#/types/git";
+import OpenHands from "#/api/open-hands";
+
+export function useUrlSearch(inputValue: string, provider: Provider) {
+  const [urlSearchResults, setUrlSearchResults] = useState<GitRepository[]>([]);
+  const [isUrlSearchLoading, setIsUrlSearchLoading] = useState(false);
+
+  useEffect(() => {
+    const handleUrlSearch = async () => {
+      if (inputValue.startsWith("https://")) {
+        const match = inputValue.match(/https:\/\/[^/]+\/([^/]+\/[^/]+)/);
+        if (match) {
+          const repoName = match[1];
+          console.log("URL detected, searching for:", repoName);
+          setIsUrlSearchLoading(true);
+          try {
+            const repositories = await OpenHands.searchGitRepositories(
+              repoName,
+              3,
+              provider,
+            );
+            console.log("URL search results:", repositories);
+            setUrlSearchResults(repositories);
+          } catch (error) {
+            console.error("URL search failed:", error);
+            setUrlSearchResults([]);
+          } finally {
+            setIsUrlSearchLoading(false);
+          }
+        }
+      } else {
+        setUrlSearchResults([]);
+      }
+    };
+
+    handleUrlSearch();
+  }, [inputValue, provider]);
+
+  return { urlSearchResults, isUrlSearchLoading };
+}
